F - Legg til ny route for å oppdatere meldekortDag R - Flytt domene klasser til egne filer R - Flytt dto til egne filer
