+++ src/main/kotlin/no/nav/tiltakspenger/meldekort/api/routes/dto/DayHasBegunDTO.kt
@@ -1,0 +1,5 @@
+package no.nav.tiltakspenger.meldekort.api.routes.dto
+
+import java.time.LocalDate
+
+data class DayHasBegunDTO(val date: LocalDate)
