+++ src/main/kotlin/no/nav/tiltakspenger/meldekort/api/routes/dto/MeldekortDagDTO.kt
@@ -1,0 +1,10 @@
+package no.nav.tiltakspenger.meldekort.api.routes.dto
+
+import java.time.LocalDate
+
+data class MeldekortDagDTO(
+    val meldekortId: String,
+    val dato: LocalDate,
+    val tiltakId: String,
+    val status: String,
+)
