+++ src/main/kotlin/no/nav/tiltakspenger/meldekort/api/routes/dto/MeldekortGrunnlagDTO.kt
@@ -1,0 +1,27 @@
+package no.nav.tiltakspenger.meldekort.api.routes.dto
+
+import java.time.LocalDate
+
+data class MeldekortGrunnlagDTO(
+    val vedtakId: String,
+    val behandlingId: String,
+    val status: StatusDTO,
+    val vurderingsperiode: PeriodeDTO,
+    val tiltak: List<TiltakDTO>,
+)
+
+enum class StatusDTO {
+    AKTIV,
+    IKKE_AKTIV,
+}
+
+data class TiltakDTO(
+    val periodeDTO: PeriodeDTO,
+    val typeBeskrivelse: String,
+    val typeKode: String,
+    val antDagerIUken: Float,
+)
+data class PeriodeDTO(
+    val fra: LocalDate,
+    val til: LocalDate,
+)
